% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBalance.R
\name{CheckBalance}
\alias{CheckBalance}
\title{Check the balance of the adjusted covariates by histogram, barplot and
qqplot}
\usage{
CheckBalance(data, ps, trt, ind)
}
\description{
Check the balance of the adjusted covariates by histogram, barplot and
qqplot
}

