% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_train.R
\name{cv_train}
\alias{cv_train}
\title{Generates a plot of cross-validated prediction error against decay values.}
\usage{
cv_train(ps_formula, data, folds = 10, decay = 2^(seq(-10, 0, 0.5)),
  size = 8, ncores = 1, MaxNWts = 2000, maxit = 2000)
}
\arguments{
\item{ps_formula}{The propensity score formula.}

\item{data}{The data set.}

\item{folds}{Number of folds for k-fold cross validation.}

\item{size}{The number of hidden layers.}

\item{ncores}{The number of processors to use.}

\item{dec_vals}{The vector of decay values to try.}

\item{num_reps}{The number of random starts to average over.}
}
\description{
Generates a plot of cross-validated prediction error against decay values.
}

