% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoSelection.R
\name{LassoSelection}
\alias{LassoSelection}
\title{Use Lasso algorithm to select predictors for propensity score estimation.}
\usage{
LassoSelection(formula, data, nfolds = 10, standardize = T,
  family = "gaussian")
}
\arguments{
\item{formula}{Lasso regression formula.}

\item{data}{Data frame for building model.}

\item{nfolds}{number of folds, default is 10.}

\item{standardize}{Standardize covariates or not.}

\item{family}{Response type. Only "gaussian" and "binomial" are supported. If the family = "gaussian", then the criterion for cross-validation is "mse", otherwise, the criterion is the area under roc curve.}
}
\description{
Use Lasso algorithm to select predictors for propensity score estimation.
}

