% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovlp.R
\name{ovlp}
\alias{ovlp}
\title{Plot the histogram for each group, examining if the observations' PS/PS logit distributions are overlapped.}
\usage{
ovlp(trt, lps)
}
\arguments{
\item{trt}{Treatment assignment}

\item{lps}{Logit propensity score}
}
\description{
Plot the histogram for each group, examining if the observations' PS/PS logit distributions are overlapped.
}

