% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovlp_ind.R
\name{ovlp_ind}
\alias{ovlp_ind}
\title{Examine if there is an overlapping for each observation.}
\usage{
ovlp_ind(trt, lps, caliper = 0.1)
}
\arguments{
\item{trt}{Treatment assignment}

\item{lps}{Logit propensity score}

\item{caliper}{Parameter defining nonoverlapping}
}
\description{
Examine if there is an overlapping for each observation.
}

